% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{HeptathlonSeoul1988_df}
\alias{HeptathlonSeoul1988_df}
\title{Olympic Heptathlon Results - Seoul 1988}
\format{
A data frame with 25 observations and 8 variables:
\describe{
  \item{hurdles}{Time in seconds for the 100m hurdles (numeric)}
  \item{highjump}{Height in meters for the high jump (numeric)}
  \item{shot}{Distance in meters for the shot put (numeric)}
  \item{run200m}{Time in seconds for the 200m run (numeric)}
  \item{longjump}{Distance in meters for the long jump (numeric)}
  \item{javelin}{Distance in meters for the javelin throw (numeric)}
  \item{run800m}{Time in seconds for the 800m run (numeric)}
  \item{score}{Total heptathlon score (integer)}
}
}
\source{
Data taken from the \pkg{HSAUR3} package version 1.0-15
}
\usage{
data(HeptathlonSeoul1988_df)
}
\description{
This dataset, HeptathlonSeoul1988_df, is a data frame containing the results of the
Olympic heptathlon competition held in Seoul in 1988. It includes performance
metrics for each of the seven events as well as the total score.
}
\details{
The dataset name has been kept as 'HeptathlonSeoul1988_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
SouthKoreAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a base R data frame object. The original content has not been modified
in any way.
}
