% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{KoreanElection2017_df}
\alias{KoreanElection2017_df}
\title{2017 Korea Presidential Election Data}
\format{
A data frame with 1250 observations and 9 variables:
\describe{
  \item{PrecinctCode}{Precinct code (integer)}
  \item{CityCode}{City code (integer)}
  \item{CandidateName}{Candidate identifier code (integer)}
  \item{AveAge}{Average age of the voting population (numeric)}
  \item{AveYearEdu}{Average years of education (numeric)}
  \item{AveHousePrice}{Average house price (numeric)}
  \item{AveInsurance}{Average insurance enrollment indicator or count (integer)}
  \item{VoteRate}{Voter turnout rate (numeric)}
  \item{NumVote}{Number of votes cast (integer)}
}
}
\source{
Data taken from the \pkg{KPC} package version 0.1.2
}
\usage{
data(KoreanElection2017_df)
}
\description{
This dataset, KoreanElection2017_df, is a data frame containing information from
the 2017 presidential election in South Korea. It includes precinct- and city-level data
along with demographic and socioeconomic indicators related to the voting population.
}
\details{
The dataset name has been kept as 'KoreanElection2017_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
SouthKoreAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame object. The original content has not been modified
in any way.
}
