% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{NFIColumnNames_df}
\alias{NFIColumnNames_df}
\title{Korean and English Column Names}
\format{
A data frame with 174 observations and 3 variables:
\describe{
  \item{Korean_Column_Name}{Column name in Korean (character)}
  \item{English_Name}{Column name in English (character)}
  \item{Column_Name}{Standardized column identifier (character)}
}
}
\source{
Data taken from the \pkg{knfi} package version 1.0.1.9
}
\usage{
data(NFIColumnNames_df)
}
\description{
This dataset, NFIColumnNames_df, is a data frame containing Korean and English translations
of column names, along with their standardized column identifiers. It is intended to assist
users in mapping between Korean-language variables and their English equivalents in related datasets.
}
\details{
The dataset name has been kept as 'NFIColumnNames_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
SouthKoreAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame object. The original content has not been modified
in any way.
}
