% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{SeoulDistrictPop_df}
\alias{SeoulDistrictPop_df}
\title{Seoul's Population and Area Data for Districts (2012)}
\format{
A data frame with 25 observations and 5 variables:
\describe{
  \item{District}{Name of the district (character)}
  \item{City}{Name of the city (character)}
  \item{Pop.2012}{Population in 2012 (integer)}
  \item{Area}{Area of the district in square kilometers (numeric)}
  \item{Founded}{Year the district was founded (character)}
}
}
\source{
Data taken from the \pkg{micromapST} package version 3.0.4
}
\usage{
data(SeoulDistrictPop_df)
}
\description{
This dataset, SeoulDistrictPop_df, is a data frame containing the 2012 population and area
data for each of the districts in the city of Seoul, South Korea. It also includes
information on the founding year of each district.
}
\details{
The dataset name has been kept as 'SeoulDistrictPop_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
SouthKoreAPIs package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a base R data frame object. The original content has not been modified
in any way.
}
