% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{SouthKoreaBirths_tbl_df}
\alias{SouthKoreaBirths_tbl_df}
\title{Births in South Korea}
\format{
A tibble with 1,872 observations and 7 variables:
\describe{
  \item{age}{Age group of the mother (character)}
  \item{region}{Region name (factor with 16 levels)}
  \item{time}{Calendar year (integer)}
  \item{births}{Number of births (integer)}
  \item{popn}{Mid-year population (integer)}
  \item{gdp_pc_2023}{GDP per capita in 2023 (numeric)}
  \item{dens_2020}{Population density in 2020 (character)}
}
}
\source{
Data taken from the \pkg{bage} package version 0.9.4
}
\usage{
data(SouthKoreaBirths_tbl_df)
}
\description{
This dataset, SouthKoreaBirths_tbl_df, is a tibble containing births and mid-year population
data for South Korea by age of mother, region, and calendar year from 2011 to 2023.
It also includes regional data on GDP per capita (2023) and population density (2020).
}
\details{
The dataset name has been kept as 'SouthKoreaBirths_tbl_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
SouthKoreAPIs package and assists users in identifying its specific characteristics.
The suffix 'tbl_df' indicates that the dataset is a tibble object. The original content has not been modified
in any way.
}
