% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southkoreapis-package.R
\docType{package}
\name{SouthKoreAPIs}
\alias{SouthKoreAPIs}
\alias{SouthKoreAPIs-package}
\title{SouthKoreAPIs: Access South Korean Data via Public APIs and Curated Datasets}
\description{
{
This package provides functions to access data from public RESTful APIs including 'Nager.Date', 'World Bank API', and 'REST Countries API', retrieving real-time or historical data related to South Korea, such as holidays, economic indicators, and international demographic and geopolitical indicators. Additionally, the package includes one of the largest curated collections of open datasets focused on South Korea, covering topics such as public health outbreaks, demographics, social surveys, elections, economic indicators, natural disasters, administrative divisions, air quality, climate data, energy consumption, cultural information, and financial markets.
}
}
\details{
SouthKoreAPIs: Access South Korean Data via Public APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access South Korean Data via Public APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/southkoreapis}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
